One of the things I find amusing is reading people's code who picked up programming as a hobby or without any formal training. It is almost invariably without design, fore-thought, or any kind of code analysis.
Logic errors, redundancies, etc. ... smh...

Arduino programmers are so adorable. This code although appears complete is poorly written I fail to see any reason to mention the original author. By the time we are done with it, it will not resemble anything like the original. But for the sake annoying academia: https://www.allaboutcircuits.com/projects/gps-transmission-with-the-hc-12-transmitter/ and https://www.arduino.cc/en/Tutorial/Datalogger